Move builtin classes into its own module. Implement `QuakeClass` for `PointLight`, `SpotLight`, and `DirectionalLight`.
